Snowpack.ODBC (Snowpack v0.5.4)

Adapter to Erlang's :odbc module.

This module is a GenServer that handles communication between Elixir and Erlang's :odbc module.

It is used by Snowpack.Protocol and should not generally be accessed directly.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

describe(pid, table)

Specs

describe(pid(), iodata()) ::
  {:selected, [binary()], [tuple()]} | {:error, Snowpack.Error.t()}

Describes the given table.

describe_result(pid, query_id)

Specs

describe_result(pid(), iodata()) ::
  {:selected, [binary()], [tuple()]} | {:error, Snowpack.Error.t()}

disconnect(pid)

Specs

disconnect(pid()) :: :ok

Disconnects from the ODBC driver.

pid is the :odbc process id

init(opts)

Specs

init(keyword()) :: {:ok, any()}

Callback implementation for GenServer.init/1.

query(pid, statement, params, opts, with_query_id \\ false)

Specs

query(pid(), iodata(), Keyword.t(), Keyword.t(), boolean()) ::
  {:selected, [binary()], [tuple()]}
  | {:selected, [binary()], [tuple()], [{binary()}]}
  | {:updated, non_neg_integer()}
  | {:error, Snowpack.Error.t()}

Sends a parametrized query to the ODBC driver.

Interface to :odbc.param_query/3.See Erlang's ODBC guide for usage details and examples.

pid is the :odbc process id

statement is the SQL query string

params are the parameters to send with the SQL query

opts are options to be passed on to :odbc

timeout in millisecods (defaults to 2 minutes)

with_query_id runs query in transaction and selects LAST_QUERY_ID()

start_link(conn_str, opts)

Specs

start_link(binary(), Keyword.t()) :: {:ok, pid()}

Starts the connection process to the ODBC driver.

conn_str should be a connection string in the format required by your ODBC driver. opts will be passed verbatim to :odbc.connect/2.